requirements.txt
ipython
jupyterlab
loguru
matplotlib
mkdocs
notebook
numpy
pandas
pip
python-dotenv
ruff
scikit-learn
statsmodels
meteostat
pyarrow
tqdm
typer
-e .
